refactor: extract edit commands + execute functions into ai-edit.ts (-1,790 lines)
Extracts 6 editing commands and their self-contained execute functions:

ai-edit.ts (~1,840 lines):
  Commands: silence-cut, caption, noise-reduce, fade, translate-srt, jump-cut
  Execute functions: executeSilenceCut, executeJumpCut, executeCaption,
    executeNoiseReduce, executeFade, executeTranslateSrt,
    applyTextOverlays, executeTextOverlay
  Types: SilencePeriod, FillerWord, CaptionStyle, TextOverlayStyle, ...

ai.ts:
  Imports applyTextOverlays, TextOverlayStyle etc. for use in pipeline cmds
  Re-exports all execute functions/types for backward compat
  Shrinks from 10,816 → 9,047 lines (-1,790 net)

All 232 tests pass.
